+++ src/Filter/TermFilter.php
@@ -2,7 +2,7 @@
 
 namespace Novaway\ElasticsearchClient\Filter;
 
-use Novaway\ElasticsearchClient\Filter\Term\TermQuery;
+use Novaway\ElasticsearchClient\Query\Term\TermQuery;
 
 /**
  * @deprecated use Novaway\ElasticsearchClient\Filter\Term\TermQuery instead
