+++ src/Query/BoolQuery.php
@@ -3,43 +3,9 @@
 
 namespace Novaway\ElasticsearchClient\Query;
 
-
-use Novaway\ElasticsearchClient\Clause;
-use Webmozart\Assert\Assert;
-
-class BoolQuery implements Query
+/**
+ * @deprecated use Novaway\ElasticsearchClient\Query\Compound\BoolQuery instead
+ */
+class BoolQuery extends \Novaway\ElasticsearchClient\Query\Compound\BoolQuery
 {
-
-    /** @var string */
-    private $combiningFactor;
-    /** @var Clause[] */
-    private $clauses;
-
-    public function __construct(string $combiningFactor = CombiningFactor::MUST)
-    {
-        Assert::oneOf($combiningFactor, CombiningFactor::toArray());
-        $this->combiningFactor = $combiningFactor;
-        $this->clauses = [];
-    }
-
-    public function getCombiningFactor(): string
-    {
-        return $this->combiningFactor;
-    }
-
-    public function addClause(Clause $clause)
-    {
-        $this->clauses[] = $clause;
-    }
-    public function formatForQuery(): array
-    {
-        $res = [];
-        foreach ($this->clauses as $clause) {
-            $res[$clause->getCombiningFactor()][] = $clause->formatForQuery();
-        }
-
-        return ['bool' => $res];
-    }
-
-
 }
